requirements.txt
python-dotenv
openai
yt-dlp
faster-whisper